Add model-provider UI, DB and token quotas
Introduce model provider management and per-flow embedding token accounting across frontend, backend and DB. Changes include: DB schema updates (new model_provider_configs table; added embedding token/chunk fields to file_upload; use BIGINT for rate limits). Frontend: new Model Provider view, router entries, locales and typings; knowledge-base shows estimated/actual embedding usage and syncs tasks from server; usage-monitor now models token-budget limits for LLM/Embedding upload/query and related UI/clone helpers; added a Playwright test and test result. Backend: refactor EmbeddingClient to use ModelProviderConfigService, support UPLOAD/QUERY usage, reserve/settle token reservations per flow, build provider-specific WebClient, and expose model version/usage result; add model provider services/repository and related tests. Overall enables configurable model providers and global token budgeting for embedding/LLM flows.

+++ src/main/java/com/yizhaoqi/smartpai/model/RateLimitConfig.java
@@ -26,13 +26,13 @@ public class RateLimitConfig {
     private Long singleWindowSeconds;
 
     @Column(name = "minute_max")
-    private Integer minuteMax;
+    private Long minuteMax;
 
     @Column(name = "minute_window_seconds")
     private Long minuteWindowSeconds;
 
     @Column(name = "day_max")
-    private Integer dayMax;
+    private Long dayMax;
 
     @Column(name = "day_window_seconds")
     private Long dayWindowSeconds;
